[WPE] Split WPEWebView into Legacy and Platform implementations https://bugs.webkit.org/show_bug.cgi?id=275985
Reviewed by Nikolas Zimmermann.

It makes clearer what code is expected to be used for new and old APIs.

* Source/WebKit/SourcesWPE.txt:
* Source/WebKit/UIProcess/API/C/wpe/WKView.cpp:
(WKViewCreate):
(WKViewCreateDeprecated):
* Source/WebKit/UIProcess/API/glib/WebKitWebView.cpp:
(webkitWebViewCreatePage):
(webkitWebViewGetRendererBufferFormat):
* Source/WebKit/UIProcess/API/wpe/PageClientImpl.cpp:
(WebKit::PageClientImpl::doneWithTouchEvent):
(WebKit::PageClientImpl::enterAcceleratedCompositingMode):
(WebKit::PageClientImpl::exitAcceleratedCompositingMode):
(WebKit::PageClientImpl::updateAcceleratedCompositingMode):
(WebKit::PageClientImpl::enterFullScreen):
(WebKit::PageClientImpl::exitFullScreen):
* Source/WebKit/UIProcess/API/wpe/WPEWebView.cpp:
(WKWPE::View::View):
(WKWPE::View::close):
(WKWPE::View::accessible const):
(): Deleted.
(WKWPE::m_backend): Deleted.
(WKWPE::View::~View): Deleted.
(WKWPE::View::setViewState): Deleted.
(WKWPE::View::handleKeyboardEvent): Deleted.
(WKWPE::View::synthesizeCompositionKeyPress): Deleted.
(WKWPE::viewToplevelIsFullScreen): Deleted.
(WKWPE::View::enterFullScreen): Deleted.
(WKWPE::View::didEnterFullScreen): Deleted.
(WKWPE::View::exitFullScreen): Deleted.
(WKWPE::View::didExitFullScreen): Deleted.
(WKWPE::View::requestExitFullScreen): Deleted.
(WKWPE::View::setFullScreen): Deleted.
(WKWPE::View::platformWebPageProxyForGamepadInput): Deleted.
(WKWPE::View::updateAcceleratedSurface): Deleted.
(WKWPE::View::renderBufferFormat const): Deleted.
(WKWPE::View::updateDisplayID): Deleted.
(WKWPE::View::touchPointsForEvent): Deleted.
(WKWPE::View::setCursor): Deleted.
(WKWPE::View::callAfterNextPresentationUpdate): Deleted.
* Source/WebKit/UIProcess/API/wpe/WPEWebView.h:
(WKWPE::View::page):
(WKWPE::View::client const):
(WKWPE::View::size const):
(WKWPE::View::backend const):
(WKWPE::View::wpeView const):
(WKWPE::View::setCursor):
(WKWPE::View::create): Deleted.
(WKWPE::View::backend): Deleted.
(WKWPE::View::touchGestureController const): Deleted.
* Source/WebKit/UIProcess/API/wpe/WPEWebViewLegacy.cpp: Added.
(WKWPE::ViewLegacy::ViewLegacy):
(WKWPE::ViewLegacy::~ViewLegacy):
(WKWPE::ViewLegacy::setViewState):
(WKWPE::ViewLegacy::handleKeyboardEvent):
(WKWPE::ViewLegacy::synthesizeCompositionKeyPress):
(WKWPE::ViewLegacy::setFullScreen):
(WKWPE::ViewLegacy::platformWebPageProxyForGamepadInput):
(WKWPE::ViewLegacy::callAfterNextPresentationUpdate):
* Source/WebKit/UIProcess/API/wpe/WPEWebViewLegacy.h: Added.
* Source/WebKit/UIProcess/API/wpe/WPEWebViewPlatform.cpp: Added.
(WKWPE::ViewPlatform::ViewPlatform):
(WKWPE::ViewPlatform::~ViewPlatform):
(WKWPE::viewToplevelIsFullScreen):
(WKWPE::ViewPlatform::enterFullScreen):
(WKWPE::ViewPlatform::didEnterFullScreen):
(WKWPE::ViewPlatform::exitFullScreen):
(WKWPE::ViewPlatform::didExitFullScreen):
(WKWPE::ViewPlatform::requestExitFullScreen):
(WKWPE::ViewPlatform::updateAcceleratedSurface):
(WKWPE::ViewPlatform::renderBufferFormat const):
(WKWPE::ViewPlatform::updateDisplayID):
(WKWPE::ViewPlatform::touchPointsForEvent):
(WKWPE::ViewPlatform::synthesizeCompositionKeyPress):
(WKWPE::ViewPlatform::setCursor):
(WKWPE::ViewPlatform::callAfterNextPresentationUpdate):
* Source/WebKit/UIProcess/API/wpe/WPEWebViewPlatform.h: Added.
* Source/WebKit/UIProcess/Gamepad/libwpe/UIGamepadProviderLibWPE.cpp:
(WebKit::UIGamepadProvider::platformWebPageProxyForGamepadInput):

Canonical link: https://commits.webkit.org/280522@main
