+++ Source/WebKit/UIProcess/API/wpe/PageClientImpl.cpp
@@ -32,7 +32,8 @@
 #include "NativeWebTouchEvent.h"
 #include "NativeWebWheelEvent.h"
 #include "TouchGestureController.h"
-#include "WPEWebView.h"
+#include "WPEWebViewLegacy.h"
+#include "WPEWebViewPlatform.h"
 #include "WebContextMenuProxy.h"
 #include "WebContextMenuProxyWPE.h"
 #include "WebKitPopupMenu.h"
@@ -213,6 +214,11 @@ void PageClientImpl::doneWithKeyEvent(const NativeWebKeyboardEvent&, bool)
 #if ENABLE(TOUCH_EVENTS)
 void PageClientImpl::doneWithTouchEvent(const NativeWebTouchEvent& touchEvent, bool wasEventHandled)
 {
+#if ENABLE(WPE_PLATFORM)
+    if (m_view.wpeView())
+        return;
+#endif
+
     if (wasEventHandled)
         return;
 
@@ -221,7 +227,7 @@ void PageClientImpl::doneWithTouchEvent(const NativeWebTouchEvent& touchEvent, b
         return;
 
     auto& page = m_view.page();
-    auto& touchGestureController = m_view.touchGestureController();
+    auto& touchGestureController = static_cast<WKWPE::ViewLegacy&>(m_view).touchGestureController();
 
     auto generatedEvent = touchGestureController.handleEvent(touchPoint);
     WTF::switchOn(generatedEvent,
@@ -275,7 +281,8 @@ Ref<WebContextMenuProxy> PageClientImpl::createContextMenuProxy(WebPageProxy& pa
 void PageClientImpl::enterAcceleratedCompositingMode(const LayerTreeContext& context)
 {
 #if ENABLE(WPE_PLATFORM)
-    m_view.updateAcceleratedSurface(context.contextID);
+    if (m_view.wpeView())
+        static_cast<WKWPE::ViewPlatform&>(m_view).updateAcceleratedSurface(context.contextID);
 #else
     UNUSED_PARAM(context);
 #endif
@@ -284,14 +291,16 @@ void PageClientImpl::enterAcceleratedCompositingMode(const LayerTreeContext& con
 void PageClientImpl::exitAcceleratedCompositingMode()
 {
 #if ENABLE(WPE_PLATFORM)
-    m_view.updateAcceleratedSurface(0);
+    if (m_view.wpeView())
+        static_cast<WKWPE::ViewPlatform&>(m_view).updateAcceleratedSurface(0);
 #endif
 }
 
 void PageClientImpl::updateAcceleratedCompositingMode(const LayerTreeContext& context)
 {
 #if ENABLE(WPE_PLATFORM)
-    m_view.updateAcceleratedSurface(context.contextID);
+    if (m_view.wpeView())
+        static_cast<WKWPE::ViewPlatform&>(m_view).updateAcceleratedSurface(context.contextID);
 #endif
 }
 
@@ -389,14 +398,14 @@ void PageClientImpl::enterFullScreen()
     m_view.willEnterFullScreen();
 #if ENABLE(WPE_PLATFORM)
     if (m_view.wpeView()) {
-        m_view.enterFullScreen();
+        static_cast<WKWPE::ViewPlatform&>(m_view).enterFullScreen();
         return;
     }
 #endif
 
     WebFullScreenManagerProxy* fullScreenManagerProxy = m_view.page().fullScreenManager();
     if (fullScreenManagerProxy) {
-        if (!m_view.setFullScreen(true))
+        if (!static_cast<WKWPE::ViewLegacy&>(m_view).setFullScreen(true))
             fullScreenManagerProxy->didExitFullScreen();
     }
 }
@@ -409,14 +418,14 @@ void PageClientImpl::exitFullScreen()
     m_view.willExitFullScreen();
 #if ENABLE(WPE_PLATFORM)
     if (m_view.wpeView()) {
-        m_view.exitFullScreen();
+        static_cast<WKWPE::ViewPlatform&>(m_view).exitFullScreen();
         return;
     }
 #endif
 
     WebFullScreenManagerProxy* fullScreenManagerProxy = m_view.page().fullScreenManager();
     if (fullScreenManagerProxy) {
-        if (!m_view.setFullScreen(false))
+        if (!static_cast<WKWPE::ViewLegacy&>(m_view).setFullScreen(false))
             fullScreenManagerProxy->didEnterFullScreen();
     }
 }
