+++ Source/WebKit/UIProcess/API/wpe/WPEWebView.h
@@ -26,11 +26,9 @@
 #pragma once
 
 #include "APIObject.h"
-#include "GRefPtrWPE.h"
 #include "InputMethodFilter.h"
 #include "KeyAutoRepeatHandler.h"
 #include "PageClientImpl.h"
-#include "RendererBufferFormat.h"
 #include "WebFullScreenManagerProxy.h"
 #include "WebKitWebViewAccessible.h"
 #include "WebPageProxy.h"
@@ -38,13 +36,9 @@
 #include <memory>
 #include <wtf/OptionSet.h>
 #include <wtf/RefPtr.h>
-#include <wtf/glib/GRefPtr.h>
 
-typedef struct OpaqueJSContext* JSGlobalContextRef;
 typedef struct _WebKitInputMethodContext WebKitInputMethodContext;
-typedef struct _WPEDisplay WPEDisplay;
 typedef struct _WPEView WPEView;
-struct wpe_input_keyboard_event;
 struct wpe_view_backend;
 
 namespace API {
@@ -57,12 +51,7 @@ struct CompositionUnderline;
 }
 
 namespace WebKit {
-class AcceleratedBackingStoreDMABuf;
-class TouchGestureController;
 class WebKitWebResourceLoadManager;
-class WebPageGroup;
-class WebPlatformTouchPoint;
-class WebProcessPool;
 struct EditingRange;
 struct UserMessage;
 }
@@ -71,19 +60,7 @@ namespace WKWPE {
 
 class View : public API::ObjectImpl<API::Object::Type::View> {
 public:
-#if ENABLE(WPE_PLATFORM)
-    static View* create(struct wpe_view_backend* backend, WPEDisplay* display, const API::PageConfiguration& configuration)
-    {
-        return new View(backend, display, configuration);
-    }
-#else
-    static View* create(struct wpe_view_backend* backend, const API::PageConfiguration& configuration)
-    {
-        return new View(backend, configuration);
-    }
-#endif
-
-    ~View();
+    virtual ~View();
 
     // Client methods
     void setClient(std::unique_ptr<API::ViewClient>&&);
@@ -96,117 +73,52 @@ public:
     void setInputMethodContext(WebKitInputMethodContext*);
     WebKitInputMethodContext* inputMethodContext() const;
     void setInputMethodState(std::optional<WebKit::InputMethodState>&&);
-    void synthesizeCompositionKeyPress(const String&, std::optional<Vector<WebCore::CompositionUnderline>>&&, std::optional<WebKit::EditingRange>&&);
-
-    void selectionDidChange();
-
-    WebKit::WebPageProxy& page() { return *m_pageProxy; }
-
-    API::ViewClient& client() const { return *m_client; }
-    struct wpe_view_backend* backend() { return m_backend; }
-#if ENABLE(WPE_PLATFORM)
-    WPEView* wpeView() const { return m_wpeView.get(); }
-#endif
-
-    const WebCore::IntSize& size() const { return m_size; }
-
-    OptionSet<WebCore::ActivityState> viewState() const { return m_viewStateFlags; }
-
-    void close();
 
 #if ENABLE(FULLSCREEN_API)
     bool isFullScreen() const;
     void willEnterFullScreen();
-#if ENABLE(WPE_PLATFORM)
-    void enterFullScreen();
-    void didEnterFullScreen();
-#endif
     void willExitFullScreen();
-#if ENABLE(WPE_PLATFORM)
-    void exitFullScreen();
-    void didExitFullScreen();
-    void requestExitFullScreen();
 #endif
-    bool setFullScreen(bool);
-#endif
+
+    void selectionDidChange();
+    void close();
+
+    WebKit::WebPageProxy& page() { return *m_pageProxy; }
+    API::ViewClient& client() const { return *m_client; }
+    const WebCore::IntSize& size() const { return m_size; }
+    OptionSet<WebCore::ActivityState> viewState() const { return m_viewStateFlags; }
 
 #if USE(ATK)
     WebKitWebViewAccessible* accessible() const;
 #endif
 
-#if ENABLE(TOUCH_EVENTS)
-    WebKit::TouchGestureController& touchGestureController() const { return *m_touchGestureController; }
-#endif
-#if ENABLE(GAMEPAD)
-    static WebKit::WebPageProxy* platformWebPageProxyForGamepadInput();
-#endif
-
+    virtual struct wpe_view_backend* backend() const { return nullptr; }
 #if ENABLE(WPE_PLATFORM)
-    void updateAcceleratedSurface(uint64_t);
-    WebKit::RendererBufferFormat renderBufferFormat() const;
+    virtual WPEView* wpeView() const { return nullptr; }
 #endif
 
-    void setCursor(const WebCore::Cursor&);
+    virtual void setCursor(const WebCore::Cursor&) { };
+    virtual void synthesizeCompositionKeyPress(const String&, std::optional<Vector<WebCore::CompositionUnderline>>&&, std::optional<WebKit::EditingRange>&&) = 0;
+    virtual void callAfterNextPresentationUpdate(CompletionHandler<void()>&&) = 0;
 
-    void callAfterNextPresentationUpdate(CompletionHandler<void()>&&);
-
-private:
-#if ENABLE(WPE_PLATFORM)
-    View(struct wpe_view_backend*, WPEDisplay*, const API::PageConfiguration&);
-#else
-    View(struct wpe_view_backend*, const API::PageConfiguration&);
-#endif
+protected:
+    explicit View(const API::PageConfiguration&);
 
     void setSize(const WebCore::IntSize&);
-    void setViewState(OptionSet<WebCore::ActivityState>);
-    void handleKeyboardEvent(struct wpe_input_keyboard_event*);
-
-#if ENABLE(WPE_PLATFORM)
-    void updateDisplayID();
-#endif
-
-#if ENABLE(TOUCH_EVENTS) && ENABLE(WPE_PLATFORM)
-    Vector<WebKit::WebPlatformTouchPoint> touchPointsForEvent(WPEEvent*);
-#endif
 
     std::unique_ptr<API::ViewClient> m_client;
-
-#if ENABLE(TOUCH_EVENTS)
-    std::unique_ptr<WebKit::TouchGestureController> m_touchGestureController;
-#if ENABLE(WPE_PLATFORM)
-    HashMap<uint32_t, GRefPtr<WPEEvent>, IntHash<uint32_t>, WTF::UnsignedWithZeroKeyHashTraits<uint32_t>> m_touchEvents;
-#endif
-#endif
     std::unique_ptr<WebKit::PageClientImpl> m_pageClient;
     RefPtr<WebKit::WebPageProxy> m_pageProxy;
     WebCore::IntSize m_size;
     OptionSet<WebCore::ActivityState> m_viewStateFlags;
-
-    struct wpe_view_backend* m_backend;
-#if ENABLE(WPE_PLATFORM)
-    GRefPtr<WPEView> m_wpeView;
-    std::unique_ptr<WebKit::AcceleratedBackingStoreDMABuf> m_backingStore;
-    uint32_t m_displayID { 0 };
-    unsigned long m_bufferRenderedID { 0 };
-    CompletionHandler<void()> m_nextPresentationUpdateCallback;
-#endif
-
 #if ENABLE(FULLSCREEN_API)
     WebKit::WebFullScreenManagerProxy::FullscreenState m_fullscreenState { WebKit::WebFullScreenManagerProxy::FullscreenState::NotInFullscreen };
-#if ENABLE(WPE_PLATFORM)
-    bool m_viewWasAlreadyInFullScreen { false };
 #endif
-#endif
-
+    WebKit::InputMethodFilter m_inputMethodFilter;
+    WebKit::KeyAutoRepeatHandler m_keyAutoRepeatHandler;
 #if USE(ATK)
     mutable GRefPtr<WebKitWebViewAccessible> m_accessible;
 #endif
-
-    bool m_horizontalScrollActive { false };
-    bool m_verticalScrollActive { false };
-
-    WebKit::InputMethodFilter m_inputMethodFilter;
-    WebKit::KeyAutoRepeatHandler m_keyAutoRepeatHandler;
 };
 
 } // namespace WKWPE
